refactor(snapshot): extract incremental update into snapshot/incremental.rs + reorganize mod.rs
Split `snapshot/mod.rs` so the incremental-update state machine (Case A/B/C/D.1/D.2/E/F)
and its tests live in a new `snapshot/incremental.rs`. Reorganize the remaining
`impl Snapshot` block in `mod.rs` into labeled sections: construction entry points,
internal constructors, field accessors and state queries, downstream operation
builders, mutations, and test-only accessors. Move the long state-machine doc from
`builder_from` (a one-liner wrapper) to `try_new_from` (the implementation).

Pure refactor: zero behavior change, all 7492 tests pass.

`commit()` test helper is hoisted to `snapshot` module scope (gated `#[cfg(test)]`)
so both `mod.rs`'s tests and `incremental.rs`'s tests share it without duplication.
