+++ kernel/src/snapshot/mod.rs
@@ -24,7 +24,6 @@ use crate::crc::{
 use crate::expressions::ColumnName;
 use crate::incremental_scan::IncrementalScanBuilder;
 use crate::log_segment::{DomainMetadataMap, LogSegment};
-use crate::log_segment_files::LogSegmentFiles;
 use crate::metrics::MetricId;
 use crate::path::ParsedLogPath;
 use crate::scan::ScanBuilder;
@@ -38,6 +37,7 @@ use crate::utils::require;
 use crate::{DeltaResult, Engine, Error, LogCompactionWriter, Version};
 
 mod builder;
+mod incremental;
 pub use builder::SnapshotBuilder;
 
 /// A shared, thread-safe reference to a [`Snapshot`].
@@ -101,6 +101,10 @@ impl std::fmt::Debug for Snapshot {
 }
 
 impl Snapshot {
+    // ============================================================================
+    // Construction entry points
+    // ============================================================================
+
     /// Create a new [`SnapshotBuilder`] to build a new [`Snapshot`] for a given table root. If you
     /// instead have an existing [`Snapshot`] you would like to do minimal work to update, consider
     /// using [`Snapshot::builder_from`] instead.
@@ -108,68 +112,17 @@ impl Snapshot {
         SnapshotBuilder::new_for(table_root)
     }
 
-    /// Create a new [`SnapshotBuilder`] to incrementally update a [`Snapshot`] to a more recent
-    /// version. Variables used below:
+    /// Create a new [`SnapshotBuilder`] to incrementally update an existing [`Snapshot`] to a
+    /// more recent version.
     ///
-    /// - `T` = effective target version: the time-travel version from
-    ///   [`SnapshotBuilder::at_version`] when set, otherwise the `max_catalog_version` from
-    ///   [`SnapshotBuilder::with_max_catalog_version`] for catalog-managed tables, otherwise unset
-    ///   (defaults to latest).
-    /// - `S1` = existing snapshot version.
-    /// - `S2` = new snapshot version (= end version of the new listing). When `T` is set, the
-    ///   listing is upper-bounded by `T`, so `S2 == T`.
-    /// - `C1` = existing checkpoint version (if any).
-    /// - `C2` = new checkpoint version (if found in the listing). Always `<= S2`, since `C2` is by
-    ///   definition a checkpoint observed within the listing. By transitivity, `C2 <= T` when `T`
-    ///   is set.
-    ///
-    /// The log listing is catalog-log-tail aware: any `log_tail` provided to the builder is
-    /// merged with filesystem listings.
-    ///
-    /// Position layout per case (`....` is the version axis; `====` marks the range read for
-    /// P+M replay; `listed` is the listing range; `read` is what's read for P+M):
-    ///
-    /// ```text
-    ///             positions                       listed       read       outcome
-    /// Case A:     C1 ........ S1 (= T)            -            -          return existing
-    /// Case B:     C1 .. T .. S1                   -            -          error (T < S1)
-    /// Case C.1:   C1 .. S1 .. T                   empty        -          error (T unreachable)
-    /// Case C.2:   C1 .. S1                        empty        -          return existing
-    /// Case D.1:   C1 .. S1 .. C2 ===== S2         [C1+1, S2]   [C2, S2]   rebuild from C2
-    /// Case D.2:   C1 . C2 .. S1 ====== S2         [C1+1, S2]   (S1, S2]   advance base, -> F
-    /// Case E:     C1 .. S1 (= S2)                 [C1+1, S1]   -          return existing
-    /// Case F:     C1 .. S1 ====== S2              [C1+1, S2]   (S1, S2]   incremental update
-    /// ```
-    ///
-    /// In the incremental cases (D.2, F), the existing snapshot's P+M at `S1` is the baseline
-    /// and only commits in `(S1, S2]` are replayed.
-    ///
-    /// - **A.** `T == S1`: return the existing snapshot unchanged.
-    /// - **B.** `T < S1`: error. The incremental path only moves forward.
-    /// - Otherwise (`T` unset or `T > S1`), list the log from `C1+1` (or from version 1 if there is
-    ///   no existing checkpoint), and one of the following applies:
-    ///   - **C.** Listing is empty:
-    ///     - **C.1.** `T` is set: error (target is newer than anything in the log).
-    ///     - **C.2.** `T` is unset: return the existing snapshot.
-    ///   - **D.** Listing contains a checkpoint:
-    ///     - **D.1.** `C2 > S1`: the new checkpoint at `C2` already captures the table state
-    ///       through version `C2`, including changes in `(S1, C2]`, so we can use it as the new
-    ///       base instead of replaying those commits. Build a fresh snapshot from `C2`.
-    ///     - **D.2.** `C2 <= S1`: the existing snapshot's P+M (at `S1`) already reflects everything
-    ///       `C2` would tell us, so we skip reading `C2` for P+M. Fall through to case F to replay
-    ///       only commits `> S1` for P+M; the combined segment uses `C2` as its checkpoint base,
-    ///       producing a better log segment with fewer deltas above the checkpoint (e.g. faster
-    ///       distributed log replay).
-    ///   - **E.** Listing contains commits but no new checkpoint, and `S2 == S1`: return the
-    ///     existing snapshot.
-    ///   - **F.** Listing contains new commits (no new checkpoint, or fall through from D.2): run
-    ///     lightweight P+M replay on commits `> S1` and merge them into the existing log segment.
-    ///
-    /// Each case is marked with `// Case X` in `Snapshot::try_new_from`. The engine is passed
-    /// to [`SnapshotBuilder::build`].
+    /// See `Snapshot::try_new_from` for the case-by-case behavior.
     pub fn builder_from(existing_snapshot: SnapshotRef) -> SnapshotBuilder {
         SnapshotBuilder::new_from(existing_snapshot)
     }
+
+    // ============================================================================
+    // Internal constructors
+    // ============================================================================
 
     /// Create a new [`Snapshot`] from a [`LogSegment`] and [`TableConfiguration`].
     #[internal_api]
@@ -201,296 +154,6 @@ impl Snapshot {
             table_configuration,
             lazy_crc,
         }
-    }
-
-    /// Create a new [`Snapshot`] instance from an existing [`Snapshot`]. This is useful when you
-    /// already have a [`Snapshot`] lying around and want to do the minimal work to 'update' the
-    /// snapshot to a later version.
-    #[instrument(err, fields(version, operation_id = %operation_id), skip(engine, target_version))]
-    fn try_new_from(
-        existing_snapshot: Arc<Snapshot>,
-        log_tail: Vec<ParsedLogPath>,
-        engine: &dyn Engine,
-        target_version: impl Into<Option<Version>>,
-        operation_id: MetricId,
-    ) -> DeltaResult<Arc<Self>> {
-        let existing_log_segment = &existing_snapshot.log_segment;
-        let existing_snapshot_version = existing_snapshot.version();
-        let requested_version = target_version.into();
-        if let Some(requested_version) = requested_version {
-            tracing::Span::current().record("version", requested_version);
-            // Case A: re-requesting the same version.
-            if requested_version == existing_snapshot_version {
-                return Ok(existing_snapshot.clone());
-            }
-            // Case B: incremental path only moves forward.
-            if requested_version < existing_snapshot_version {
-                return Err(Error::Generic(format!(
-                    "Requested snapshot version {requested_version} is older than snapshot \
-                    hint version {existing_snapshot_version}"
-                )));
-            }
-        } else {
-            tracing::Span::current().record("version", existing_snapshot_version);
-        }
-
-        let log_root = existing_log_segment.log_root.clone();
-        let storage = engine.storage_handler();
-
-        // Start listing just after the previous segment's checkpoint, if any.
-        let listing_start = existing_log_segment.checkpoint_version.unwrap_or(0) + 1;
-
-        // Check for new commits (and CRC)
-        let new_listed_files = LogSegmentFiles::list(
-            storage.as_ref(),
-            &log_root,
-            log_tail,
-            Some(listing_start),
-            requested_version,
-        )?;
-
-        // NB: we need to check both checkpoints and commits since we filter commits at and below
-        // the checkpoint version. Example: if we have a checkpoint + commit at version 1, the log
-        // listing above will only return the checkpoint and not the commit.
-        if new_listed_files.ascending_commit_files().is_empty()
-            && new_listed_files.checkpoint_parts().is_empty()
-        {
-            match requested_version {
-                // Case C.1: caller requested a specific version (necessarily >
-                // existing_snapshot_version since cases A and B were handled above), but
-                // no such commit exists in the log.
-                Some(requested_version) => {
-                    return Err(Error::Generic(format!(
-                        "Requested snapshot version {requested_version} is not available: \
-                         no new commits were found after existing snapshot version \
-                         {existing_snapshot_version}"
-                    )));
-                }
-                // Case C.2: no new commits and no explicit target; latest is existing.
-                None => {
-                    return Ok(existing_snapshot.clone());
-                }
-            }
-        }
-
-        // create a log segment just from existing_checkpoint.version -> new_version
-        // OR could be from 1 -> new_version
-        // Save the latest_commit before moving new_listed_files
-        let new_latest_commit_file = new_listed_files.latest_commit_file().clone();
-        // Note: new_log_segment won't have last_checkpoint_metadata since we're listing without a
-        // hint. If the new segment has a checkpoint, we will return it as is. Otherwise, we
-        // will preserve last_checkpoint_metadata when merging the new log segment with the
-        // old one.
-        let mut new_log_segment =
-            LogSegment::try_new(new_listed_files, log_root.clone(), requested_version, None)?;
-
-        let new_end_version = new_log_segment.end_version;
-        if new_end_version < existing_snapshot_version {
-            // we should never see a new log segment with a version < the existing snapshot
-            // version, that would mean a commit was incorrectly deleted from the log
-            return Err(Error::Generic(format!(
-                "Unexpected state: the newest version in the log {new_end_version} is \
-                 older than the existing snapshot version {existing_snapshot_version}"
-            )));
-        }
-        if let Some(new_checkpoint_version) = new_log_segment.checkpoint_version {
-            if new_checkpoint_version > existing_snapshot_version {
-                // Case D.1: checkpoint ahead of existing snapshot. Commits between
-                // existing_snapshot_version+1 and new_checkpoint_version were filtered out
-                // of the listing, so a full rebuild is required.
-                // `resolve_crc` reuses the existing `LazyCrc` when the resolved CRC version
-                // matches, avoiding redundant I/O.
-                let (_, lazy_crc) = Self::resolve_crc(
-                    &new_log_segment,
-                    existing_log_segment,
-                    &existing_snapshot.lazy_crc,
-                );
-                let snapshot = Self::try_new_from_log_segment(
-                    existing_snapshot.table_root().clone(),
-                    new_log_segment,
-                    engine,
-                    operation_id,
-                    Some(lazy_crc),
-                );
-                return Ok(Arc::new(snapshot?));
-            }
-            // Case D.2: checkpoint at or below existing snapshot; fall through to case F to
-            // replay only the new commits and advance the checkpoint base.
-        }
-
-        // Case E: no new checkpoint, version did not advance; return existing.
-        if new_end_version == existing_snapshot_version
-            && new_log_segment.checkpoint_version.is_none()
-        {
-            // We must check checkpoint_version here: if a checkpoint at or below the existing
-            // snapshot version was discovered, we still need to fall through to advance the
-            // checkpoint base even though the version did not change.
-            return Ok(existing_snapshot.clone());
-        }
-
-        // Case F: lightweight P+M replay on new commits, merge with existing segment.
-        // (Also reached from Case D.2 when a checkpoint at or below the existing snapshot
-        // version was discovered.)
-        //
-        // The example below illustrates Case D.2 -> F.
-        #[rustfmt::skip]
-        // Example: existing segment = checkpoint@v0 + commit@v1, v2, v3
-        //          existing_snapshot_version = v3, listing_start = v1
-        //          target=v4, new checkpoint@v2 written since last build (Case D.2 -> F):
-        //
-        //    listing:          commit@v1, then checkpoint@v2 + commit@v3, v4
-        //                      (checkpoint flush drops v1)
-        //    new segment:      checkpoint@v2 + commit@v3, v4
-        //    after dedup:      drop commit@v3 (already in existing snapshot)
-        //                      -> ascending_commit_files = [commit@v4]
-        //    commit@v3 is not lost: the existing segment contributes it below
-        //    (existing commits above checkpoint@v2 = [commit@v3])
-        //    combined segment: checkpoint@v2 + commit@v3 (from existing) + commit@v4 (from new)
-        //    checkpoint@v2 is kept in checkpoint_parts to advance the listing base.
-        new_log_segment
-            .listed
-            .ascending_commit_files
-            .retain(|log_path| existing_snapshot_version < log_path.version);
-        // Deduplicate compaction files the same way: the new listing re-lists from
-        // checkpoint_version, so it includes compaction files already in the existing segment.
-        // Note: This removes all _new_ compaction files that start at or before
-        // `existing_snapshot_version`, which may drop useful compaction files that span
-        // across the old/new boundary (e.g. a new compaction(1, 3) when
-        // existing_snapshot_version=2). This is conservative but safe.
-        new_log_segment
-            .listed
-            .ascending_compaction_files
-            .retain(|log_path| existing_snapshot_version < log_path.version);
-
-        let (crc_file, lazy_crc) = Self::resolve_crc(
-            &new_log_segment,
-            existing_log_segment,
-            &existing_snapshot.lazy_crc,
-        );
-
-        // Replay only the new commits (> existing_snapshot_version) for P&M, excluding the
-        // checkpoint. Skip a stale CRC (at or below existing_snapshot_version): if no P&M is
-        // found in the pruned replay, read_protocol_metadata_opt falls back to the CRC's P&M,
-        // which could predate changes already in the existing snapshot.
-        let (new_metadata, new_protocol) = {
-            let no_crc = LazyCrc::new(None);
-            let crc: &LazyCrc = if lazy_crc
-                .crc_version()
-                .is_some_and(|v| v <= existing_snapshot_version)
-            {
-                &no_crc
-            } else {
-                &lazy_crc
-            };
-            new_log_segment
-                .segment_after_crc(existing_snapshot_version)
-                .read_protocol_metadata_opt(engine, crc)?
-        };
-        let table_configuration = TableConfiguration::try_new_from(
-            existing_snapshot.table_configuration(),
-            new_metadata,
-            new_protocol,
-            new_log_segment.end_version,
-        )?;
-
-        // Combine existing and new log segments. When the new listing found a checkpoint at or
-        // below the existing snapshot version, trim existing commits and compaction files
-        // already covered by it; otherwise keep all existing files.
-        let new_checkpoint_version = new_log_segment.checkpoint_version;
-        let new_checkpoint_hint = new_log_segment.last_checkpoint_hint_summary();
-        let mut ascending_commit_files = Self::files_after_checkpoint(
-            &existing_log_segment.listed.ascending_commit_files,
-            new_checkpoint_version,
-        );
-        ascending_commit_files.extend(new_log_segment.listed.ascending_commit_files);
-        let mut ascending_compaction_files = Self::files_after_checkpoint(
-            &existing_log_segment.listed.ascending_compaction_files,
-            new_checkpoint_version,
-        );
-        ascending_compaction_files.extend(new_log_segment.listed.ascending_compaction_files);
-
-        // When the new listing found a checkpoint at or below the existing snapshot version,
-        // advance the base so future listings start from new_checkpoint_version+1; otherwise
-        // preserve the existing base.
-        // Note: the incremental listing never uses a _last_checkpoint hint, so
-        // new_log_segment.last_checkpoint_hint_summary() is always None when a new checkpoint
-        // is found here. The combined segment will read the new checkpoint's parquet footer at
-        // scan time. When no new checkpoint was found, preserve the existing schema hint if
-        // available.
-        let (new_checkpoint_parts, new_checkpoint_schema) = if new_checkpoint_version.is_some() {
-            (
-                new_log_segment.listed.checkpoint_parts.clone(),
-                new_checkpoint_hint,
-            )
-        } else {
-            (
-                existing_log_segment.listed.checkpoint_parts.clone(),
-                existing_log_segment.last_checkpoint_hint_summary(),
-            )
-        };
-        let combined_log_segment = LogSegment::try_new(
-            LogSegmentFiles {
-                ascending_commit_files,
-                ascending_compaction_files,
-                checkpoint_parts: new_checkpoint_parts,
-                latest_crc_file: crc_file,
-                // In Case F there are new commits (new_end_version > existing_snapshot_version
-                // with no new checkpoint), so the new listing's `latest_commit_file` is always
-                // `Some(commit @ new_end_version)` and matches the combined snapshot version.
-                latest_commit_file: new_latest_commit_file,
-                max_published_version: new_log_segment
-                    .listed
-                    .max_published_version
-                    .max(existing_log_segment.listed.max_published_version),
-            },
-            log_root,
-            requested_version,
-            new_checkpoint_schema,
-        )?;
-
-        tracing::Span::current().record("version", table_configuration.version());
-        Ok(Arc::new(Snapshot::new_with_crc(
-            combined_log_segment,
-            table_configuration,
-            lazy_crc,
-        )))
-    }
-
-    /// Determine the CRC file and LazyCrc for an incremental snapshot update.
-    ///
-    /// Prefers the new segment's CRC file; falls back to the existing segment's CRC if and
-    /// only if its version is >= the new segment's checkpoint version. If the resolved CRC
-    /// version matches the existing snapshot's LazyCrc version, reuses it to avoid redundant
-    /// I/O (it may already be loaded in memory).
-    ///
-    /// The filter preserves the [`LogSegmentFiles`] invariant `crc.version >=
-    /// checkpoint.version`. Without it, a stale inherited CRC could silently produce wrong
-    /// P&M via [`LogSegment::read_protocol_metadata_opt`]'s Case 2(b) fallback (see test
-    /// `test_incremental_snapshot_drops_stale_crc_preserves_correct_metadata`).
-    fn resolve_crc(
-        new_log_segment: &LogSegment,
-        existing_log_segment: &LogSegment,
-        existing_lazy_crc: &Arc<LazyCrc>,
-    ) -> (Option<ParsedLogPath>, Arc<LazyCrc>) {
-        let new_crc_file = new_log_segment.listed.latest_crc_file.clone();
-        let crc_satisfies_new_ckpt = |crc: &ParsedLogPath| {
-            new_log_segment
-                .checkpoint_version
-                .is_none_or(|ckpt_v| crc.version >= ckpt_v)
-        };
-        let existing_crc_file = existing_log_segment
-            .listed
-            .latest_crc_file
-            .clone()
-            .filter(crc_satisfies_new_ckpt);
-        let crc_file = new_crc_file.or(existing_crc_file);
-        let crc_version = crc_file.as_ref().map(|f| f.version);
-        let lazy_crc = if crc_version == existing_lazy_crc.crc_version() {
-            existing_lazy_crc.clone()
-        } else {
-            Arc::new(LazyCrc::new(crc_file.clone()))
-        };
-        (crc_file, lazy_crc)
     }
 
     /// Create a new [`Snapshot`] instance.
@@ -596,177 +259,9 @@ impl Snapshot {
         }
     }
 
-    /// Creates a [`CheckpointWriter`] for generating a checkpoint from this snapshot.
-    ///
-    /// See the [`crate::checkpoint`] module documentation for more details on checkpoint types
-    /// and the overall checkpoint process.
-    pub fn create_checkpoint_writer(self: Arc<Self>) -> DeltaResult<CheckpointWriter> {
-        CheckpointWriter::try_new(self)
-    }
-
-    /// Performs a complete checkpoint of this snapshot using the provided engine.
-    ///
-    /// If a checkpoint already exists at this version, returns
-    /// [`CheckpointWriteResult::AlreadyExists`] with the original snapshot unchanged.
-    /// Otherwise, writes a checkpoint parquet file and the `_last_checkpoint` file and returns
-    /// [`CheckpointWriteResult::Written`] with an updated [`SnapshotRef`] whose log segment
-    /// reflects the new checkpoint. Commits and compaction files subsumed by the checkpoint are
-    /// dropped from the returned snapshot.
-    ///
-    /// # Parameters
-    /// - `engine`: Engine for data processing and I/O
-    /// - `spec`: Checkpoint format specification. `None` uses the default checkpoint settings
-    ///   (auto-detecting V1/V2 from table features). For V2 checkpoints, the default is to not
-    ///   write sidecar files.
-    ///
-    /// # Errors
-    /// - If `CheckpointSpec::V2` is used but the table does not support the `v2Checkpoint` feature.
-    /// - If `CheckpointSpec::V1` is used but the table supports `v2Checkpoint` feature. Note: the
-    ///   Delta protocol permits writing V1 checkpoints to such tables; this is a kernel limitation.
-    /// - If `file_actions_per_sidecar_hint` is `Some(0)`.
-    /// - If the checkpoint write fails (e.g. I/O, parquet write). A `FileAlreadyExists` error is
-    ///   not propagated; it returns [`CheckpointWriteResult::AlreadyExists`] instead. Note: this
-    ///   also fires on the (unlikely) case of a sidecar UUID filename collision, where
-    ///   it should ideally surface as an error. Tracked in
-    ///   <https://github.com/delta-io/delta-kernel-rs/issues/2503>.
-    ///
-    /// Note:
-    ///     - It is still possible that an existing checkpoint gets overwritten if that checkpoint
-    ///       was written by a concurrent writer.
-    ///     - This function uses [`crate::ParquetHandler::write_parquet_file`] and
-    ///       [`crate::StorageHandler::head`], which may not be implemented by all engines. If you
-    ///       are using the default engine, make sure to build it with the multi-threaded executor
-    ///       if you want to use this method.
-    ///
-    /// [`CheckpointSpec`]: crate::checkpoint::CheckpointSpec
-    ///
-    /// Note: There is currently no public api for callers to determine whether a table supports V2
-    /// checkpoints directly. Tracked in <https://github.com/delta-io/delta-kernel-rs/issues/2450>.
-    #[instrument(parent = &self.span, name = "snap.checkpoint", skip_all, err)]
-    pub fn checkpoint(
-        self: &SnapshotRef,
-        engine: &dyn Engine,
-        spec: Option<&CheckpointSpec>,
-    ) -> DeltaResult<(CheckpointWriteResult, SnapshotRef)> {
-        if self.log_segment.checkpoint_version == Some(self.log_segment.end_version) {
-            info!(
-                "Checkpoint already exists for snapshot version {}",
-                self.version()
-            );
-            return Ok((CheckpointWriteResult::AlreadyExists, Arc::clone(self)));
-        }
-        info!(
-            "Writing checkpoint for snapshot version {} with spec {:?}",
-            self.version(),
-            spec
-        );
-
-        let v2_supported = self
-            .table_configuration()
-            .is_feature_supported(&TableFeature::V2Checkpoint);
-        match spec {
-            Some(CheckpointSpec::V2(cfg)) => {
-                if !v2_supported {
-                    return Err(Error::checkpoint_write(
-                        "CheckpointSpec::V2 requires the v2Checkpoint table feature to be supported",
-                    ));
-                }
-                if let V2CheckpointConfig::WithSidecar {
-                    file_actions_per_sidecar_hint: Some(0),
-                } = cfg
-                {
-                    return Err(Error::checkpoint_write(
-                        "file_actions_per_sidecar_hint must be greater than 0",
-                    ));
-                }
-            }
-            Some(CheckpointSpec::V1) if v2_supported => {
-                // TODO: remove this once we support writing V1 checkpoints even if table supports
-                // v2Checkpoint See <https://github.com/delta-io/delta-kernel-rs/issues/2454>.
-                return Err(Error::unsupported(
-                    "Kernel does not support writing V1 checkpoints when the table supports v2Checkpoint",
-                ));
-            }
-            _ => {}
-        }
-
-        let writer = Arc::clone(self).create_checkpoint_writer()?;
-
-        let write_result = match spec {
-            Some(CheckpointSpec::V2(V2CheckpointConfig::WithSidecar {
-                file_actions_per_sidecar_hint,
-            })) => {
-                let hint =
-                    file_actions_per_sidecar_hint.unwrap_or(DEFAULT_FILE_ACTIONS_PER_SIDECAR_HINT);
-                writer.write_v2_checkpoint_with_sidecars(engine, hint)
-            }
-            _ => writer.write_checkpoint_without_sidecars(engine),
-        };
-
-        let info = match write_result {
-            Ok(info) => info,
-            Err(Error::FileAlreadyExists(_)) => {
-                // NOTE: Per write_parquet_file's documentation, it should silently overwrite
-                // existing files, so we log a warning but still return the correct result.
-                warn!(
-                    "ParquetHandler::write_parquet_file unexpectedly failed on \
-                    FileAlreadyExists for version {}",
-                    self.version()
-                );
-                return Ok((CheckpointWriteResult::AlreadyExists, Arc::clone(self)));
-            }
-            Err(e) => return Err(e),
-        };
-
-        writer.finalize(engine, &info.last_checkpoint_stats)?;
-
-        let checkpoint_log_path = ParsedLogPath::try_from(info.file_meta)?.ok_or_else(|| {
-            Error::internal_error("Checkpoint path could not be parsed as a log path")
-        })?;
-        let checkpoint_version = checkpoint_log_path.version;
-        let new_log_segment = self
-            .log_segment
-            .try_new_with_checkpoint(checkpoint_log_path)?;
-        // Drop a cached CRC that predates the checkpoint version.
-        let lazy_crc = if self
-            .lazy_crc
-            .crc_version()
-            .is_some_and(|v| v < checkpoint_version)
-        {
-            Arc::new(LazyCrc::new(None))
-        } else {
-            self.lazy_crc.clone()
-        };
-        Ok((
-            CheckpointWriteResult::Written,
-            Arc::new(Snapshot::new_with_crc(
-                new_log_segment,
-                self.table_configuration().clone(),
-                lazy_crc,
-            )),
-        ))
-    }
-
-    /// Creates a [`LogCompactionWriter`] for generating a log compaction file.
-    ///
-    /// Log compaction aggregates commit files in a version range into a single compacted file,
-    /// improving performance by reducing the number of files to process during log replay.
-    ///
-    /// # Parameters
-    /// - `start_version`: The first version to include in the compaction (inclusive)
-    /// - `end_version`: The last version to include in the compaction (inclusive)
-    ///
-    /// # Returns
-    /// A [`LogCompactionWriter`] that can be used to generate the compaction file.
-    ///
-    /// NOTE: This method is currently a no-op because log compaction is disabled (#2337)
-    pub fn log_compaction_writer(
-        self: Arc<Self>,
-        start_version: Version,
-        end_version: Version,
-    ) -> DeltaResult<LogCompactionWriter> {
-        LogCompactionWriter::try_new(self, start_version, end_version)
-    }
+    // ============================================================================
+    // Field accessors and state queries
+    // ============================================================================
 
     /// Log segment this snapshot uses
     #[internal_api]
@@ -847,45 +342,6 @@ impl Snapshot {
     #[internal_api]
     pub(crate) fn table_configuration(&self) -> &TableConfiguration {
         &self.table_configuration
-    }
-
-    /// Create a [`ScanBuilder`] for an `SnapshotRef`.
-    pub fn scan_builder(self: Arc<Self>) -> ScanBuilder {
-        ScanBuilder::new(self)
-    }
-
-    /// Create an [`IncrementalScanBuilder`] for the range `(base_version, self.version()]`.
-    ///
-    /// Use this to advance a cached file listing from `base_version` to this snapshot's
-    /// version without doing a full scan. See [`IncrementalScanBuilder`] for details.
-    pub fn incremental_scan_builder(
-        self: Arc<Self>,
-        base_version: Version,
-    ) -> IncrementalScanBuilder {
-        IncrementalScanBuilder::new(self, base_version)
-    }
-
-    /// Create a [`Transaction`] for this `SnapshotRef`. With the specified [`Committer`].
-    ///
-    /// Note: For tables with clustering enabled, this performs log replay to read clustering
-    /// columns from domain metadata, which may have a performance cost.
-    pub fn transaction(
-        self: Arc<Self>,
-        committer: Box<dyn Committer>,
-        engine: &dyn Engine,
-    ) -> DeltaResult<Transaction> {
-        Transaction::try_new_existing_table(self, committer, engine)
-    }
-
-    /// Creates a builder for altering this table's metadata. Currently supports schema change
-    /// operations.
-    ///
-    /// The returned builder allows chaining operations before building an
-    /// [`AlterTableTransaction`] that can be committed.
-    ///
-    /// [`AlterTableTransaction`]: crate::transaction::AlterTableTransaction
-    pub fn alter_table(self: Arc<Self>) -> AlterTableTransactionBuilder {
-        AlterTableTransactionBuilder::new(self)
     }
 
     /// Fetch the latest version of the provided `application_id` for this snapshot. Filters the
@@ -1078,185 +534,6 @@ impl Snapshot {
         self.log_segment().scan_domain_metadatas(domains, engine)
     }
 
-    /// Returns file-level statistics, or `None` if no CRC at this snapshot's version has
-    /// `Complete` file stats. Attempts to load the CRC from storage if not already cached.
-    pub fn get_or_load_file_stats(&self, engine: &dyn Engine) -> Option<FileStats> {
-        let crc = self
-            .lazy_crc
-            .get_or_load_if_at_version(engine, self.version())?;
-        crc.file_stats().cloned()
-    }
-
-    /// Returns file-level statistics if the CRC is already loaded at this snapshot's version.
-    /// This method performs no I/O; it returns `None` if the CRC has not already been loaded.
-    pub fn get_file_stats_if_loaded(&self) -> Option<FileStats> {
-        let crc = self.lazy_crc.get_if_loaded_at_version(self.version())?;
-        crc.file_stats().cloned()
-    }
-
-    /// Returns the CRC if one has been loaded at this snapshot's version (no I/O).
-    ///
-    /// This is a test-only helper for integration tests to inspect the CRC state.
-    #[cfg(any(test, feature = "test-utils"))]
-    pub fn get_current_crc_if_loaded_for_testing(&self) -> Option<&Crc> {
-        if self.lazy_crc.crc_version() != Some(self.version()) {
-            return None;
-        }
-        self.lazy_crc.cached.get()?.get().map(|arc| arc.as_ref())
-    }
-
-    /// Returns the CRC version tracked by this snapshot's LazyCrc, if any.
-    ///
-    /// This is a test-only helper for integration tests to inspect the CRC version.
-    #[cfg(any(test, feature = "test-utils"))]
-    pub fn crc_version_for_testing(&self) -> Option<Version> {
-        self.lazy_crc.crc_version()
-    }
-
-    /// Writes a version checksum (CRC) file for this snapshot. Writers should call this after
-    /// every commit because checksums enable faster snapshot loading and table state validation.
-    ///
-    /// Currently only supports writing from a post-commit snapshot that has pre-computed CRC
-    /// information in memory (i.e. the snapshot returned by
-    /// [`CommittedTransaction::post_commit_snapshot`]).
-    ///
-    /// Returns a tuple of [`ChecksumWriteResult`] and a [`SnapshotRef`]. On
-    /// [`ChecksumWriteResult::Written`], the returned snapshot has the CRC file recorded in
-    /// its log segment. On [`ChecksumWriteResult::AlreadyExists`], the original snapshot is
-    /// returned unchanged.
-    ///
-    /// # Errors
-    ///
-    /// - [`Error::ChecksumWriteUnsupported`] if no in-memory CRC is available at this snapshot's
-    ///   version (e.g. a snapshot loaded from disk that has no CRC file), if the CRC's
-    ///   `file_stats_state` is `Indeterminate` (a non-incremental operation like ANALYZE STATS was
-    ///   encountered, or a file action had a missing size; recoverable with a full state
-    ///   reconstruction in the future), or if `delta.enableInCommitTimestamps` is `true` but
-    ///   `inCommitTimestampOpt` is absent.
-    /// - I/O errors from the engine's storage handler if the write fails.
-    ///
-    /// [`CommittedTransaction::post_commit_snapshot`]: crate::transaction::CommittedTransaction::post_commit_snapshot
-    #[instrument(parent = &self.span, name = "snap.write_checksum", skip_all, err)]
-    pub fn write_checksum(
-        self: &SnapshotRef,
-        engine: &dyn Engine,
-    ) -> DeltaResult<(ChecksumWriteResult, SnapshotRef)> {
-        let has_crc_on_disk = self
-            .log_segment
-            .listed
-            .latest_crc_file
-            .as_ref()
-            .is_some_and(|f| f.version == self.version());
-
-        if has_crc_on_disk {
-            info!(
-                "CRC file already exists on disk at version {}",
-                self.version()
-            );
-            return Ok((ChecksumWriteResult::AlreadyExists, Arc::clone(self)));
-        }
-
-        let crc = self
-            .lazy_crc
-            .get_if_loaded_at_version(self.version())
-            .ok_or_else(|| {
-                Error::ChecksumWriteUnsupported(
-                    "No in-memory CRC available at this snapshot version.".to_string(),
-                )
-            })?;
-
-        let crc_path = ParsedLogPath::new_crc(self.table_root(), self.version())?;
-
-        match try_write_crc_file(engine, &crc_path.location, crc) {
-            Ok(()) => {
-                info!("Wrote CRC file at {}", crc_path.location);
-                let new_log_segment = self.log_segment.try_new_with_crc_file(crc_path)?;
-                let new_snapshot = Arc::new(Snapshot::new_with_crc(
-                    new_log_segment,
-                    self.table_configuration().clone(),
-                    self.lazy_crc.clone(),
-                ));
-                Ok((ChecksumWriteResult::Written, new_snapshot))
-            }
-            Err(Error::FileAlreadyExists(_)) => {
-                info!(
-                    "Another writer beat us to writing CRC file at {}",
-                    crc_path.location
-                );
-                Ok((ChecksumWriteResult::AlreadyExists, Arc::clone(self)))
-            }
-            Err(e) => Err(e),
-        }
-    }
-
-    /// Publishes all catalog commits at this table version. Applicable only to catalog-managed
-    /// tables. This method is a no-op for filesystem-managed tables or if there are no catalog
-    /// commits to publish.
-    ///
-    /// Publishing copies ratified catalog commits to the Delta log as published Delta files,
-    /// reducing catalog storage requirements and enabling some table maintenance operations,
-    /// like checkpointing.
-    ///
-    /// # Parameters
-    ///
-    /// - `engine`: The engine to use for publishing commits
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the publish operation fails, or if there are catalog commits that need
-    /// publishing but the table or committer don't support publishing.
-    ///
-    /// # See Also
-    ///
-    /// - [`Committer::publish`]
-    #[instrument(parent = &self.span, name = "snap.publish", skip_all, err)]
-    pub fn publish(
-        self: &SnapshotRef,
-        engine: &dyn Engine,
-        committer: &dyn Committer,
-    ) -> DeltaResult<SnapshotRef> {
-        let unpublished_catalog_commits = self.log_segment().get_unpublished_catalog_commits()?;
-
-        if unpublished_catalog_commits.is_empty() {
-            return Ok(Arc::clone(self));
-        }
-
-        require!(
-            unpublished_catalog_commits
-                .windows(2)
-                .all(|commits| commits[0].version() + 1 == commits[1].version()),
-            Error::generic(format!(
-                "Expected ordered and contiguous unpublished catalog commits. \
-                 Got: {unpublished_catalog_commits:?}"
-            ))
-        );
-
-        require!(
-            self.table_configuration().is_catalog_managed(),
-            Error::generic(
-                "There are catalog commits that need publishing, but the table is not catalog-managed.",
-            )
-        );
-
-        require!(
-            committer.is_catalog_committer(),
-            Error::generic(
-                "There are catalog commits that need publishing, but the committer is not a catalog committer.",
-            )
-        );
-
-        let publish_metadata =
-            PublishMetadata::try_new(self.version(), unpublished_catalog_commits)?;
-
-        committer.publish(engine, publish_metadata)?;
-
-        Ok(Arc::new(Snapshot::new_with_crc(
-            self.log_segment().new_as_published()?,
-            self.table_configuration().clone(),
-            self.lazy_crc.clone(),
-        )))
-    }
-
     /// Fetch both user-controlled and system-controlled domain metadata for a specific domain
     /// in this snapshot.
     ///
@@ -1288,6 +565,22 @@ impl Snapshot {
             .into_values()
             .filter(|domain| !domain.is_internal())
             .collect())
+    }
+
+    /// Returns file-level statistics, or `None` if no CRC at this snapshot's version has
+    /// `Complete` file stats. Attempts to load the CRC from storage if not already cached.
+    pub fn get_or_load_file_stats(&self, engine: &dyn Engine) -> Option<FileStats> {
+        let crc = self
+            .lazy_crc
+            .get_or_load_if_at_version(engine, self.version())?;
+        crc.file_stats().cloned()
+    }
+
+    /// Returns file-level statistics if the CRC is already loaded at this snapshot's version.
+    /// This method performs no I/O; it returns `None` if the CRC has not already been loaded.
+    pub fn get_file_stats_if_loaded(&self) -> Option<FileStats> {
+        let crc = self.lazy_crc.get_if_loaded_at_version(self.version())?;
+        crc.file_stats().cloned()
     }
 
     /// Get the In-Commit Timestamp (ICT) for this snapshot.
@@ -1402,21 +695,409 @@ impl Snapshot {
         }
     }
 
-    /// Filters `files` to only those with version above `checkpoint_version`, or clones all if
-    /// `checkpoint_version` is `None`.
-    fn files_after_checkpoint(
-        files: &[ParsedLogPath],
-        checkpoint_version: Option<Version>,
-    ) -> Vec<ParsedLogPath> {
-        match checkpoint_version {
-            Some(ckpt_v) => files
-                .iter()
-                .filter(|f| f.version > ckpt_v)
-                .cloned()
-                .collect(),
-            None => files.to_vec(),
+    // ============================================================================
+    // Downstream operation builders
+    // ============================================================================
+
+    /// Create a [`ScanBuilder`] for an `SnapshotRef`.
+    pub fn scan_builder(self: Arc<Self>) -> ScanBuilder {
+        ScanBuilder::new(self)
+    }
+
+    /// Create an [`IncrementalScanBuilder`] for the range `(base_version, self.version()]`.
+    ///
+    /// Use this to advance a cached file listing from `base_version` to this snapshot's
+    /// version without doing a full scan. See [`IncrementalScanBuilder`] for details.
+    pub fn incremental_scan_builder(
+        self: Arc<Self>,
+        base_version: Version,
+    ) -> IncrementalScanBuilder {
+        IncrementalScanBuilder::new(self, base_version)
+    }
+
+    /// Create a [`Transaction`] for this `SnapshotRef`. With the specified [`Committer`].
+    ///
+    /// Note: For tables with clustering enabled, this performs log replay to read clustering
+    /// columns from domain metadata, which may have a performance cost.
+    pub fn transaction(
+        self: Arc<Self>,
+        committer: Box<dyn Committer>,
+        engine: &dyn Engine,
+    ) -> DeltaResult<Transaction> {
+        Transaction::try_new_existing_table(self, committer, engine)
+    }
+
+    /// Creates a builder for altering this table's metadata. Currently supports schema change
+    /// operations.
+    ///
+    /// The returned builder allows chaining operations before building an
+    /// [`AlterTableTransaction`] that can be committed.
+    ///
+    /// [`AlterTableTransaction`]: crate::transaction::AlterTableTransaction
+    pub fn alter_table(self: Arc<Self>) -> AlterTableTransactionBuilder {
+        AlterTableTransactionBuilder::new(self)
+    }
+
+    /// Creates a [`CheckpointWriter`] for generating a checkpoint from this snapshot.
+    ///
+    /// See the [`crate::checkpoint`] module documentation for more details on checkpoint types
+    /// and the overall checkpoint process.
+    pub fn create_checkpoint_writer(self: Arc<Self>) -> DeltaResult<CheckpointWriter> {
+        CheckpointWriter::try_new(self)
+    }
+
+    /// Creates a [`LogCompactionWriter`] for generating a log compaction file.
+    ///
+    /// Log compaction aggregates commit files in a version range into a single compacted file,
+    /// improving performance by reducing the number of files to process during log replay.
+    ///
+    /// # Parameters
+    /// - `start_version`: The first version to include in the compaction (inclusive)
+    /// - `end_version`: The last version to include in the compaction (inclusive)
+    ///
+    /// # Returns
+    /// A [`LogCompactionWriter`] that can be used to generate the compaction file.
+    ///
+    /// NOTE: This method is currently a no-op because log compaction is disabled (#2337)
+    pub fn log_compaction_writer(
+        self: Arc<Self>,
+        start_version: Version,
+        end_version: Version,
+    ) -> DeltaResult<LogCompactionWriter> {
+        LogCompactionWriter::try_new(self, start_version, end_version)
+    }
+
+    // ============================================================================
+    // Mutations
+    // ============================================================================
+
+    /// Writes a version checksum (CRC) file for this snapshot. Writers should call this after
+    /// every commit because checksums enable faster snapshot loading and table state validation.
+    ///
+    /// Currently only supports writing from a post-commit snapshot that has pre-computed CRC
+    /// information in memory (i.e. the snapshot returned by
+    /// [`CommittedTransaction::post_commit_snapshot`]).
+    ///
+    /// Returns a tuple of [`ChecksumWriteResult`] and a [`SnapshotRef`]. On
+    /// [`ChecksumWriteResult::Written`], the returned snapshot has the CRC file recorded in
+    /// its log segment. On [`ChecksumWriteResult::AlreadyExists`], the original snapshot is
+    /// returned unchanged.
+    ///
+    /// # Errors
+    ///
+    /// - [`Error::ChecksumWriteUnsupported`] if no in-memory CRC is available at this snapshot's
+    ///   version (e.g. a snapshot loaded from disk that has no CRC file), if the CRC's
+    ///   `file_stats_state` is `Indeterminate` (a non-incremental operation like ANALYZE STATS was
+    ///   encountered, or a file action had a missing size; recoverable with a full state
+    ///   reconstruction in the future), or if `delta.enableInCommitTimestamps` is `true` but
+    ///   `inCommitTimestampOpt` is absent.
+    /// - I/O errors from the engine's storage handler if the write fails.
+    ///
+    /// [`CommittedTransaction::post_commit_snapshot`]: crate::transaction::CommittedTransaction::post_commit_snapshot
+    #[instrument(parent = &self.span, name = "snap.write_checksum", skip_all, err)]
+    pub fn write_checksum(
+        self: &SnapshotRef,
+        engine: &dyn Engine,
+    ) -> DeltaResult<(ChecksumWriteResult, SnapshotRef)> {
+        let has_crc_on_disk = self
+            .log_segment
+            .listed
+            .latest_crc_file
+            .as_ref()
+            .is_some_and(|f| f.version == self.version());
+
+        if has_crc_on_disk {
+            info!(
+                "CRC file already exists on disk at version {}",
+                self.version()
+            );
+            return Ok((ChecksumWriteResult::AlreadyExists, Arc::clone(self)));
+        }
+
+        let crc = self
+            .lazy_crc
+            .get_if_loaded_at_version(self.version())
+            .ok_or_else(|| {
+                Error::ChecksumWriteUnsupported(
+                    "No in-memory CRC available at this snapshot version.".to_string(),
+                )
+            })?;
+
+        let crc_path = ParsedLogPath::new_crc(self.table_root(), self.version())?;
+
+        match try_write_crc_file(engine, &crc_path.location, crc) {
+            Ok(()) => {
+                info!("Wrote CRC file at {}", crc_path.location);
+                let new_log_segment = self.log_segment.try_new_with_crc_file(crc_path)?;
+                let new_snapshot = Arc::new(Snapshot::new_with_crc(
+                    new_log_segment,
+                    self.table_configuration().clone(),
+                    self.lazy_crc.clone(),
+                ));
+                Ok((ChecksumWriteResult::Written, new_snapshot))
+            }
+            Err(Error::FileAlreadyExists(_)) => {
+                info!(
+                    "Another writer beat us to writing CRC file at {}",
+                    crc_path.location
+                );
+                Ok((ChecksumWriteResult::AlreadyExists, Arc::clone(self)))
+            }
+            Err(e) => Err(e),
         }
     }
+
+    /// Performs a complete checkpoint of this snapshot using the provided engine.
+    ///
+    /// If a checkpoint already exists at this version, returns
+    /// [`CheckpointWriteResult::AlreadyExists`] with the original snapshot unchanged.
+    /// Otherwise, writes a checkpoint parquet file and the `_last_checkpoint` file and returns
+    /// [`CheckpointWriteResult::Written`] with an updated [`SnapshotRef`] whose log segment
+    /// reflects the new checkpoint. Commits and compaction files subsumed by the checkpoint are
+    /// dropped from the returned snapshot.
+    ///
+    /// # Parameters
+    /// - `engine`: Engine for data processing and I/O
+    /// - `spec`: Checkpoint format specification. `None` uses the default checkpoint settings
+    ///   (auto-detecting V1/V2 from table features). For V2 checkpoints, the default is to not
+    ///   write sidecar files.
+    ///
+    /// # Errors
+    /// - If `CheckpointSpec::V2` is used but the table does not support the `v2Checkpoint` feature.
+    /// - If `CheckpointSpec::V1` is used but the table supports `v2Checkpoint` feature. Note: the
+    ///   Delta protocol permits writing V1 checkpoints to such tables; this is a kernel limitation.
+    /// - If `file_actions_per_sidecar_hint` is `Some(0)`.
+    /// - If the checkpoint write fails (e.g. I/O, parquet write). A `FileAlreadyExists` error is
+    ///   not propagated; it returns [`CheckpointWriteResult::AlreadyExists`] instead. Note: this
+    ///   also fires on the (unlikely) case of a sidecar UUID filename collision, where
+    ///   it should ideally surface as an error. Tracked in
+    ///   <https://github.com/delta-io/delta-kernel-rs/issues/2503>.
+    ///
+    /// Note:
+    ///     - It is still possible that an existing checkpoint gets overwritten if that checkpoint
+    ///       was written by a concurrent writer.
+    ///     - This function uses [`crate::ParquetHandler::write_parquet_file`] and
+    ///       [`crate::StorageHandler::head`], which may not be implemented by all engines. If you
+    ///       are using the default engine, make sure to build it with the multi-threaded executor
+    ///       if you want to use this method.
+    ///
+    /// [`CheckpointSpec`]: crate::checkpoint::CheckpointSpec
+    ///
+    /// Note: There is currently no public api for callers to determine whether a table supports V2
+    /// checkpoints directly. Tracked in <https://github.com/delta-io/delta-kernel-rs/issues/2450>.
+    #[instrument(parent = &self.span, name = "snap.checkpoint", skip_all, err)]
+    pub fn checkpoint(
+        self: &SnapshotRef,
+        engine: &dyn Engine,
+        spec: Option<&CheckpointSpec>,
+    ) -> DeltaResult<(CheckpointWriteResult, SnapshotRef)> {
+        if self.log_segment.checkpoint_version == Some(self.log_segment.end_version) {
+            info!(
+                "Checkpoint already exists for snapshot version {}",
+                self.version()
+            );
+            return Ok((CheckpointWriteResult::AlreadyExists, Arc::clone(self)));
+        }
+        info!(
+            "Writing checkpoint for snapshot version {} with spec {:?}",
+            self.version(),
+            spec
+        );
+
+        let v2_supported = self
+            .table_configuration()
+            .is_feature_supported(&TableFeature::V2Checkpoint);
+        match spec {
+            Some(CheckpointSpec::V2(cfg)) => {
+                if !v2_supported {
+                    return Err(Error::checkpoint_write(
+                        "CheckpointSpec::V2 requires the v2Checkpoint table feature to be supported",
+                    ));
+                }
+                if let V2CheckpointConfig::WithSidecar {
+                    file_actions_per_sidecar_hint: Some(0),
+                } = cfg
+                {
+                    return Err(Error::checkpoint_write(
+                        "file_actions_per_sidecar_hint must be greater than 0",
+                    ));
+                }
+            }
+            Some(CheckpointSpec::V1) if v2_supported => {
+                // TODO: remove this once we support writing V1 checkpoints even if table supports
+                // v2Checkpoint See <https://github.com/delta-io/delta-kernel-rs/issues/2454>.
+                return Err(Error::unsupported(
+                    "Kernel does not support writing V1 checkpoints when the table supports v2Checkpoint",
+                ));
+            }
+            _ => {}
+        }
+
+        let writer = Arc::clone(self).create_checkpoint_writer()?;
+
+        let write_result = match spec {
+            Some(CheckpointSpec::V2(V2CheckpointConfig::WithSidecar {
+                file_actions_per_sidecar_hint,
+            })) => {
+                let hint =
+                    file_actions_per_sidecar_hint.unwrap_or(DEFAULT_FILE_ACTIONS_PER_SIDECAR_HINT);
+                writer.write_v2_checkpoint_with_sidecars(engine, hint)
+            }
+            _ => writer.write_checkpoint_without_sidecars(engine),
+        };
+
+        let info = match write_result {
+            Ok(info) => info,
+            Err(Error::FileAlreadyExists(_)) => {
+                // NOTE: Per write_parquet_file's documentation, it should silently overwrite
+                // existing files, so we log a warning but still return the correct result.
+                warn!(
+                    "ParquetHandler::write_parquet_file unexpectedly failed on \
+                    FileAlreadyExists for version {}",
+                    self.version()
+                );
+                return Ok((CheckpointWriteResult::AlreadyExists, Arc::clone(self)));
+            }
+            Err(e) => return Err(e),
+        };
+
+        writer.finalize(engine, &info.last_checkpoint_stats)?;
+
+        let checkpoint_log_path = ParsedLogPath::try_from(info.file_meta)?.ok_or_else(|| {
+            Error::internal_error("Checkpoint path could not be parsed as a log path")
+        })?;
+        let checkpoint_version = checkpoint_log_path.version;
+        let new_log_segment = self
+            .log_segment
+            .try_new_with_checkpoint(checkpoint_log_path)?;
+        // Drop a cached CRC that predates the checkpoint version.
+        let lazy_crc = if self
+            .lazy_crc
+            .crc_version()
+            .is_some_and(|v| v < checkpoint_version)
+        {
+            Arc::new(LazyCrc::new(None))
+        } else {
+            self.lazy_crc.clone()
+        };
+        Ok((
+            CheckpointWriteResult::Written,
+            Arc::new(Snapshot::new_with_crc(
+                new_log_segment,
+                self.table_configuration().clone(),
+                lazy_crc,
+            )),
+        ))
+    }
+
+    /// Publishes all catalog commits at this table version. Applicable only to catalog-managed
+    /// tables. This method is a no-op for filesystem-managed tables or if there are no catalog
+    /// commits to publish.
+    ///
+    /// Publishing copies ratified catalog commits to the Delta log as published Delta files,
+    /// reducing catalog storage requirements and enabling some table maintenance operations,
+    /// like checkpointing.
+    ///
+    /// # Parameters
+    ///
+    /// - `engine`: The engine to use for publishing commits
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if the publish operation fails, or if there are catalog commits that need
+    /// publishing but the table or committer don't support publishing.
+    ///
+    /// # See Also
+    ///
+    /// - [`Committer::publish`]
+    #[instrument(parent = &self.span, name = "snap.publish", skip_all, err)]
+    pub fn publish(
+        self: &SnapshotRef,
+        engine: &dyn Engine,
+        committer: &dyn Committer,
+    ) -> DeltaResult<SnapshotRef> {
+        let unpublished_catalog_commits = self.log_segment().get_unpublished_catalog_commits()?;
+
+        if unpublished_catalog_commits.is_empty() {
+            return Ok(Arc::clone(self));
+        }
+
+        require!(
+            unpublished_catalog_commits
+                .windows(2)
+                .all(|commits| commits[0].version() + 1 == commits[1].version()),
+            Error::generic(format!(
+                "Expected ordered and contiguous unpublished catalog commits. \
+                 Got: {unpublished_catalog_commits:?}"
+            ))
+        );
+
+        require!(
+            self.table_configuration().is_catalog_managed(),
+            Error::generic(
+                "There are catalog commits that need publishing, but the table is not catalog-managed.",
+            )
+        );
+
+        require!(
+            committer.is_catalog_committer(),
+            Error::generic(
+                "There are catalog commits that need publishing, but the committer is not a catalog committer.",
+            )
+        );
+
+        let publish_metadata =
+            PublishMetadata::try_new(self.version(), unpublished_catalog_commits)?;
+
+        committer.publish(engine, publish_metadata)?;
+
+        Ok(Arc::new(Snapshot::new_with_crc(
+            self.log_segment().new_as_published()?,
+            self.table_configuration().clone(),
+            self.lazy_crc.clone(),
+        )))
+    }
+
+    // ============================================================================
+    // Test-only accessors
+    // ============================================================================
+
+    /// Returns the CRC if one has been loaded at this snapshot's version (no I/O).
+    ///
+    /// This is a test-only helper for integration tests to inspect the CRC state.
+    #[cfg(any(test, feature = "test-utils"))]
+    pub fn get_current_crc_if_loaded_for_testing(&self) -> Option<&Crc> {
+        if self.lazy_crc.crc_version() != Some(self.version()) {
+            return None;
+        }
+        self.lazy_crc.cached.get()?.get().map(|arc| arc.as_ref())
+    }
+
+    /// Returns the CRC version tracked by this snapshot's LazyCrc, if any.
+    ///
+    /// This is a test-only helper for integration tests to inspect the CRC version.
+    #[cfg(any(test, feature = "test-utils"))]
+    pub fn crc_version_for_testing(&self) -> Option<Version> {
+        self.lazy_crc.crc_version()
+    }
+}
+
+// TODO: unify this and lots of stuff in LogSegment tests and test_utils.
+#[cfg(test)]
+async fn commit(
+    table_root: impl AsRef<str>,
+    store: &crate::object_store::memory::InMemory,
+    version: Version,
+    commit: Vec<serde_json::Value>,
+) {
+    let commit_data = commit
+        .iter()
+        .map(ToString::to_string)
+        .collect::<Vec<String>>()
+        .join("\n");
+    test_utils::add_commit(table_root, store, version, commit_data)
+        .await
+        .unwrap();
 }
 
 #[cfg(test)]
@@ -1426,10 +1107,10 @@ mod tests {
 
     use rstest::rstest;
     use serde_json::json;
-    use test_utils::table_builder::{FeatureSet, LogState, TestTableBuilder, VersionTarget};
+    use test_utils::table_builder::{FeatureSet, LogState, VersionTarget};
     use test_utils::{add_commit, delta_path_for_version};
 
-    use super::*;
+    use super::{commit, *};
     use crate::actions::{DomainMetadata, Protocol};
     use crate::arrow::array::StringArray;
     use crate::arrow::record_batch::RecordBatch;
@@ -1443,7 +1124,7 @@ mod tests {
     use crate::object_store::path::Path;
     use crate::object_store::ObjectStoreExt as _;
     use crate::parquet::arrow::ArrowWriter;
-    use crate::path::{LogPathFileType, ParsedLogPath};
+    use crate::path::ParsedLogPath;
     use crate::schema::{DataType, StructField, StructType};
     use crate::table_features::{
         TABLE_FEATURES_MIN_READER_VERSION, TABLE_FEATURES_MIN_WRITER_VERSION,
@@ -1599,23 +1280,6 @@ mod tests {
         let schema_string = r#"{"type":"struct","fields":[{"name":"value","type":"integer","nullable":true,"metadata":{}}]}"#;
         let expected: SchemaRef = serde_json::from_str(schema_string).unwrap();
         assert_eq!(snapshot.schema(), expected);
-    }
-
-    // TODO: unify this and lots of stuff in LogSegment tests and test_utils.
-    async fn commit(
-        table_root: impl AsRef<str>,
-        store: &InMemory,
-        version: Version,
-        commit: Vec<serde_json::Value>,
-    ) {
-        let commit_data = commit
-            .iter()
-            .map(ToString::to_string)
-            .collect::<Vec<String>>()
-            .join("\n");
-        add_commit(table_root, store, version, commit_data)
-            .await
-            .unwrap();
     }
 
     // interesting cases for testing Snapshot::new_from:
@@ -2660,173 +2324,6 @@ mod tests {
     }
 
     #[test]
-    fn test_try_new_from_empty_log_tail() -> DeltaResult<()> {
-        let table = TestTableBuilder::new().build().unwrap();
-        let engine = SyncEngine::new_with_store(table.store().clone());
-
-        let base_snapshot = Snapshot::builder_for(table.table_root())
-            .at_version(0)
-            .build(&engine)?;
-
-        let result = Snapshot::try_new_from(
-            base_snapshot.clone(),
-            vec![],
-            &engine,
-            None,
-            MetricId::default(),
-        )?;
-        assert_eq!(result, base_snapshot);
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_try_new_from_latest_commit_preservation() -> DeltaResult<()> {
-        let store = Arc::new(InMemory::new());
-        let url = Url::parse("memory:///")?;
-        let engine = SyncEngine::new_with_store(store.clone());
-
-        // Create commits 0-2
-        let base_commit = vec![
-            json!({"protocol": {"minReaderVersion": 1, "minWriterVersion": 2}}),
-            json!({
-                "metaData": {
-                    "id": "test-id",
-                    "format": {"provider": "parquet", "options": {}},
-                    "schemaString": "{\"type\":\"struct\",\"fields\":[]}",
-                    "partitionColumns": [],
-                    "configuration": {},
-                    "createdTime": 1587968585495i64
-                }
-            }),
-        ];
-
-        commit(url.as_str(), store.as_ref(), 0, base_commit.clone()).await;
-        commit(
-            url.as_str(),
-            store.as_ref(),
-            1,
-            vec![json!({"commitInfo": {"timestamp": 1234}})],
-        )
-        .await;
-        commit(
-            url.as_str(),
-            store.as_ref(),
-            2,
-            vec![json!({"commitInfo": {"timestamp": 5678}})],
-        )
-        .await;
-
-        let base_snapshot = Snapshot::builder_for(url.as_str())
-            .at_version(1)
-            .build(&engine)?;
-
-        // Verify base snapshot has latest_commit_file at version 1
-        assert_eq!(
-            base_snapshot
-                .log_segment
-                .listed
-                .latest_commit_file
-                .as_ref()
-                .map(|f| f.version),
-            Some(1)
-        );
-
-        // Create log_tail with FileMeta for version 2
-        let commit_2_url = url.join("_delta_log/")?.join("00000000000000000002.json")?;
-        let file_meta = crate::FileMeta {
-            location: commit_2_url,
-            last_modified: 1234567890,
-            size: 100,
-        };
-        let parsed_path = ParsedLogPath::try_from(file_meta)?
-            .ok_or_else(|| Error::Generic("Failed to parse log path".to_string()))?;
-        let log_tail = vec![parsed_path];
-
-        // Create new snapshot from base to version 2 using try_new_from directly
-        let new_snapshot = Snapshot::try_new_from(
-            base_snapshot.clone(),
-            log_tail,
-            &engine,
-            Some(2),
-            MetricId::default(),
-        )?;
-
-        // Latest commit should now be version 2
-        assert_eq!(
-            new_snapshot
-                .log_segment
-                .listed
-                .latest_commit_file
-                .as_ref()
-                .map(|f| f.version),
-            Some(2)
-        );
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn test_try_new_from_version_boundary_cases() -> DeltaResult<()> {
-        let store = Arc::new(InMemory::new());
-        let table_root = "memory:///test_table/";
-        let engine = SyncEngine::new_with_store(store.clone());
-
-        // Create commits
-        let base_commit = vec![
-            json!({"protocol": {"minReaderVersion": 1, "minWriterVersion": 2}}),
-            json!({
-                "metaData": {
-                    "id": "test-id",
-                    "format": {"provider": "parquet", "options": {}},
-                    "schemaString": "{\"type\":\"struct\",\"fields\":[]}",
-                    "partitionColumns": [],
-                    "configuration": {},
-                    "createdTime": 1587968585495i64
-                }
-            }),
-        ];
-
-        commit(table_root, store.as_ref(), 0, base_commit).await;
-        commit(
-            table_root,
-            store.as_ref(),
-            1,
-            vec![json!({"commitInfo": {"timestamp": 1234}})],
-        )
-        .await;
-
-        let base_snapshot = Snapshot::builder_for(table_root)
-            .at_version(1)
-            .build(&engine)?;
-
-        // Test requesting same version - should return same snapshot
-        let same_version = Snapshot::try_new_from(
-            base_snapshot.clone(),
-            vec![],
-            &engine,
-            Some(1),
-            MetricId::default(),
-        )?;
-        assert!(Arc::ptr_eq(&same_version, &base_snapshot));
-
-        // Test requesting older version - should error
-        let older_version = Snapshot::try_new_from(
-            base_snapshot.clone(),
-            vec![],
-            &engine,
-            Some(0),
-            MetricId::default(),
-        );
-        assert!(matches!(
-            older_version,
-            Err(Error::Generic(msg)) if msg.contains("older than snapshot hint version")
-        ));
-
-        Ok(())
-    }
-
-    #[test]
     fn test_new_post_commit_simple() {
         // GIVEN
         let path = std::fs::canonicalize(PathBuf::from("./tests/data/basic_partitioned/")).unwrap();
@@ -2844,837 +2341,6 @@ mod tests {
         // THEN
         assert_eq!(post_commit_snapshot.version(), next_version);
         assert_eq!(post_commit_snapshot.log_segment().end_version, next_version);
-    }
-
-    // Action builders for incremental-snapshot tests. Centralized so commit setup stays
-    // consistent across tests (e.g. the schema matches `make_test_crc_json`).
-    fn protocol_action(min_reader: u32, min_writer: u32) -> serde_json::Value {
-        json!({"protocol": {"minReaderVersion": min_reader, "minWriterVersion": min_writer}})
-    }
-
-    fn metadata_action(configuration: serde_json::Value) -> serde_json::Value {
-        json!({
-            "metaData": {
-                "id": "test-id",
-                "format": {"provider": "parquet", "options": {}},
-                "schemaString": "{\"type\":\"struct\",\"fields\":[{\"name\":\"id\",\"type\":\"integer\",\"nullable\":true,\"metadata\":{}}]}",
-                "partitionColumns": [],
-                "configuration": configuration,
-                "createdTime": 1587968585495i64,
-            }
-        })
-    }
-
-    fn add_action(path: &str) -> serde_json::Value {
-        json!({
-            "add": {
-                "path": path,
-                "partitionValues": {},
-                "size": 100,
-                "modificationTime": 1000,
-                "dataChange": true,
-            }
-        })
-    }
-
-    // Helper: create a minimal test table with commits 0..num_commits.
-    async fn setup_test_table_with_commits(
-        table_root: impl AsRef<str>,
-        store: &InMemory,
-        num_commits: u64,
-    ) -> DeltaResult<()> {
-        // Commit 0: protocol + metadata + first file.
-        commit(
-            table_root.as_ref(),
-            store,
-            0,
-            vec![
-                protocol_action(1, 2),
-                metadata_action(json!({})),
-                add_action("file1.parquet"),
-            ],
-        )
-        .await;
-
-        // Additional commits with just add actions.
-        for i in 1..num_commits {
-            let path = format!("file{}.parquet", i + 1);
-            commit(table_root.as_ref(), store, i, vec![add_action(&path)]).await;
-        }
-        Ok(())
-    }
-
-    // Helper: write a compaction file
-    async fn write_compaction_file(store: &InMemory, start: u64, end: u64) -> DeltaResult<()> {
-        let content = r#"{"protocol":{"minReaderVersion":1,"minWriterVersion":2}}"#;
-        store
-            .put(
-                &test_utils::compacted_log_path_for_versions(start, end, "json"),
-                content.into(),
-            )
-            .await?;
-        Ok(())
-    }
-
-    struct IncrementalSnapshotTestContext {
-        store: Arc<InMemory>,
-        url: Url,
-        engine: Arc<SyncEngine>,
-    }
-
-    fn setup_incremental_snapshot_test() -> DeltaResult<IncrementalSnapshotTestContext> {
-        let store = Arc::new(InMemory::new());
-        let url = Url::parse("memory:///")?;
-        let engine = Arc::new(SyncEngine::new_with_store(store.clone()));
-
-        Ok(IncrementalSnapshotTestContext { store, url, engine })
-    }
-
-    /// Compares two Snapshots field-by-field. LogSegment fields are compared individually,
-    /// intentionally skipping `last_checkpoint_metadata` which is only populated on the
-    /// from-scratch path (via `_last_checkpoint` hint) and not on the incremental update path.
-    fn compare_snapshots(left: &Snapshot, right: &Snapshot) {
-        assert_eq!(left.table_configuration, right.table_configuration);
-        assert_eq!(left.log_segment.end_version, right.log_segment.end_version);
-        assert_eq!(
-            left.log_segment.checkpoint_version,
-            right.log_segment.checkpoint_version
-        );
-        assert_eq!(left.log_segment.log_root, right.log_segment.log_root);
-        assert_eq!(left.log_segment.listed, right.log_segment.listed);
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_incremental_snapshot_picks_up_checkpoint_written_at_current_version(
-    ) -> DeltaResult<()> {
-        let ctx = setup_incremental_snapshot_test()?;
-
-        setup_test_table_with_commits(ctx.url.as_str(), &ctx.store, 2).await?;
-
-        let snapshot_v1 = Snapshot::builder_for(ctx.url.as_str())
-            .at_version(1)
-            .build(ctx.engine.as_ref())?;
-        assert_eq!(snapshot_v1.log_segment.checkpoint_version, None);
-
-        snapshot_v1.clone().checkpoint(ctx.engine.as_ref(), None)?;
-
-        let fresh = Snapshot::builder_for(ctx.url.as_str()).build(ctx.engine.as_ref())?;
-        assert_eq!(fresh.version(), 1);
-        assert_eq!(fresh.log_segment.checkpoint_version, Some(1));
-
-        let updated = Snapshot::builder_from(snapshot_v1).build(ctx.engine.as_ref())?;
-        assert_eq!(updated.version(), 1);
-        assert_eq!(updated.log_segment.checkpoint_version, Some(1));
-        compare_snapshots(&updated, &fresh);
-
-        Ok(())
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_incremental_snapshot_picks_up_newer_checkpoint_below_current_version(
-    ) -> DeltaResult<()> {
-        let ctx = setup_incremental_snapshot_test()?;
-
-        setup_test_table_with_commits(ctx.url.as_str(), &ctx.store, 4).await?;
-
-        Snapshot::builder_for(ctx.url.as_str())
-            .at_version(1)
-            .build(ctx.engine.as_ref())?
-            .checkpoint(ctx.engine.as_ref(), None)?;
-
-        let snapshot_v3 = Snapshot::builder_for(ctx.url.as_str())
-            .at_version(3)
-            .build(ctx.engine.as_ref())?;
-        assert_eq!(snapshot_v3.log_segment.checkpoint_version, Some(1));
-
-        Snapshot::builder_for(ctx.url.as_str())
-            .at_version(2)
-            .build(ctx.engine.as_ref())?
-            .checkpoint(ctx.engine.as_ref(), None)?;
-
-        let fresh = Snapshot::builder_for(ctx.url.as_str()).build(ctx.engine.as_ref())?;
-        assert_eq!(fresh.version(), 3);
-        assert_eq!(fresh.log_segment.checkpoint_version, Some(2));
-
-        let updated = Snapshot::builder_from(snapshot_v3).build(ctx.engine.as_ref())?;
-        assert_eq!(updated.version(), 3);
-        assert_eq!(updated.log_segment.checkpoint_version, Some(2));
-        compare_snapshots(&updated, &fresh);
-
-        Ok(())
-    }
-
-    // CRC JSON for the standard test table (see `setup_test_table_with_commits`).
-    fn make_test_crc_json(table_size_bytes: i64, num_files: i64) -> serde_json::Value {
-        json!({
-            "tableSizeBytes": table_size_bytes,
-            "numFiles": num_files,
-            "numMetadata": 1,
-            "numProtocol": 1,
-            "protocol": {"minReaderVersion": 1, "minWriterVersion": 2},
-            "metadata": {
-                "id": "test-id",
-                "format": {"provider": "parquet", "options": {}},
-                "schemaString": "{\"type\":\"struct\",\"fields\":[{\"name\":\"id\",\"type\":\"integer\",\"nullable\":true,\"metadata\":{}}]}",
-                "partitionColumns": [],
-                "configuration": {},
-                "createdTime": 1587968585495i64
-            }
-        })
-    }
-
-    // `resolve_crc` when the new listing discovers a checkpoint *at or below* the existing
-    // snapshot version must correctly decide between reusing the existing snapshot's CRC,
-    // creating a fresh LazyCrc, or dropping the CRC entirely. Each case below sets up
-    // commits 0-3, writes an existing CRC, builds snapshot_v3, writes a checkpoint,
-    // optionally writes a new CRC, then runs the incremental update and checks the outcome.
-    // (The checkpoint-*ahead* path is covered separately by
-    // `test_incremental_snapshot_multi_hop_replay_then_rebuild_drops_stale_crc`.)
-    //
-    // Version axis for `refresh_to_newer` (existing_crc_v=2, new_ckpt_v=1, new_crc_v=3):
-    //
-    //   v0 --------- v1 --------- v2 --------- v3
-    //                new_ckpt      existing     newly-listed
-    //                              crc (loaded) crc
-    //                              ^^^^^^^^     ^^^^^^^^^^^^
-    //                         snapshot_v3's    wins on version mismatch
-    //                         lazy_crc         -> fresh unloaded LazyCrc
-    //
-    // Cases:
-    //   reuse_existing:   existing crc@v2 stays when new ckpt@v1 is below it (invariant OK)
-    //                     and no newer CRC was listed; LazyCrc Arc is reused (loaded).
-    //   refresh_to_newer: existing crc@v2 is superseded by a newly-listed crc@v3; new CRC
-    //                     version wins and produces a fresh unloaded LazyCrc.
-    //   crc_equals_ckpt:  existing crc@v2 and new ckpt@v2 sit at the same version; sanity
-    //                     coverage for the invariant boundary `crc.version >= ckpt.version`.
-    //                     (Note: does not pin `>=` vs `>` in `resolve_crc`'s filter because
-    //                     the listing re-surfaces the existing CRC, so `new_crc_file =
-    //                     Some(v2)` wins via `.or(existing)` regardless of the filter. A
-    //                     scenario where only the fallback exercises the boundary is
-    //                     structurally unreachable: it would require the existing CRC to be
-    //                     below listing_start, but the LogSegmentFiles invariant forces
-    //                     existing_crc_version >= existing_checkpoint_version = listing_start
-    //                     - 1.)
-    //   drop_below_ckpt:  existing crc@v1 would violate the LogSegmentFiles invariant with
-    //                     new ckpt@v2 (1 < 2), so it is filtered out; result has no CRC.
-    #[rstest]
-    #[case::reuse_existing(2, 1, None, Some(2), true)]
-    #[case::refresh_to_newer(2, 1, Some(3), Some(3), false)]
-    #[case::crc_equals_ckpt(2, 2, None, Some(2), true)]
-    #[case::drop_below_ckpt(1, 2, None, None, false)]
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_incremental_snapshot_crc_resolution_when_checkpoint_at_or_below_snapshot(
-        #[case] existing_crc_v: u64,
-        #[case] new_ckpt_v: u64,
-        #[case] newly_listed_crc_v: Option<u64>,
-        #[case] expected_crc_v: Option<u64>,
-        #[case] expect_loaded: bool,
-    ) -> DeltaResult<()> {
-        let ctx = setup_incremental_snapshot_test()?;
-        setup_test_table_with_commits(ctx.url.as_str(), &ctx.store, 4).await?;
-
-        ctx.store
-            .put(
-                &delta_path_for_version(existing_crc_v, "crc"),
-                make_test_crc_json(300, 3).to_string().into(),
-            )
-            .await?;
-
-        let snapshot_v3 = Snapshot::builder_for(ctx.url.as_str())
-            .at_version(3)
-            .build(ctx.engine.as_ref())?;
-        assert_eq!(snapshot_v3.lazy_crc.crc_version(), Some(existing_crc_v));
-        // Building the snapshot loaded the CRC while resolving P&M: commits after
-        // existing_crc_v contain no P&M changes, so read_protocol_metadata_opt's Case 2(b)
-        // fallback loads the CRC.
-        assert!(snapshot_v3.lazy_crc.is_loaded());
-
-        // Write the new checkpoint, then optionally plant a newer CRC at `newly_listed_crc_v`.
-        Snapshot::builder_for(ctx.url.as_str())
-            .at_version(new_ckpt_v)
-            .build(ctx.engine.as_ref())?
-            .checkpoint(ctx.engine.as_ref(), None)?;
-        if let Some(v) = newly_listed_crc_v {
-            ctx.store
-                .put(
-                    &delta_path_for_version(v, "crc"),
-                    make_test_crc_json(400, 4).to_string().into(),
-                )
-                .await?;
-        }
-
-        let updated = Snapshot::builder_from(snapshot_v3).build(ctx.engine.as_ref())?;
-
-        // Observable equivalence: incremental update matches a fresh rebuild, including
-        // `log_segment.listed.latest_crc_file` (so a filter-bypass regression is caught by
-        // `compare_snapshots`, not just by the explicit CRC assertions below).
-        let fresh = Snapshot::builder_for(ctx.url.as_str()).build(ctx.engine.as_ref())?;
-        compare_snapshots(&updated, &fresh);
-        assert_eq!(updated.version(), 3);
-        assert_eq!(updated.log_segment.checkpoint_version, Some(new_ckpt_v));
-        assert_eq!(updated.lazy_crc.crc_version(), expected_crc_v);
-        assert_eq!(
-            updated
-                .log_segment
-                .listed
-                .latest_crc_file
-                .as_ref()
-                .map(|f| f.version),
-            expected_crc_v
-        );
-        assert_eq!(updated.lazy_crc.is_loaded(), expect_loaded);
-
-        Ok(())
-    }
-
-    // Stronger regression for the `resolve_crc` invariant filter: set up a metadata change
-    // at the checkpoint version so that a stale inherited CRC would produce wrong
-    // Metadata, not just wrong bookkeeping. Without the filter, the incremental update
-    // would return Metadata from commit 0 (via Case 2(b) fallback on a below-checkpoint
-    // CRC); the fresh rebuild would return Metadata from commit 2. `compare_snapshots`
-    // catches that on `table_configuration`.
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_incremental_snapshot_drops_stale_crc_preserves_correct_metadata(
-    ) -> DeltaResult<()> {
-        let ctx = setup_incremental_snapshot_test()?;
-        let table_root = ctx.url.as_str();
-
-        // commit 0: initial P&M with empty configuration.
-        commit(
-            table_root,
-            &ctx.store,
-            0,
-            vec![
-                protocol_action(1, 2),
-                metadata_action(json!({})),
-                add_action("f0.parquet"),
-            ],
-        )
-        .await;
-        // commit 1: data-only add.
-        commit(table_root, &ctx.store, 1, vec![add_action("f1.parquet")]).await;
-        // commit 2: metadata change (new configuration property). A stale CRC@v1 would miss
-        // this.
-        commit(
-            table_root,
-            &ctx.store,
-            2,
-            vec![
-                metadata_action(json!({"delta.myKey": "myValue"})),
-                add_action("f2.parquet"),
-            ],
-        )
-        .await;
-        // commit 3: data-only add.
-        commit(table_root, &ctx.store, 3, vec![add_action("f3.parquet")]).await;
-
-        // CRC@v1 captures the BEFORE metadata (no "delta.myKey" property). This is the
-        // stale CRC we expect the filter to reject.
-        ctx.store
-            .put(
-                &delta_path_for_version(1, "crc"),
-                make_test_crc_json(200, 2).to_string().into(),
-            )
-            .await?;
-
-        let snapshot_v3 = Snapshot::builder_for(table_root)
-            .at_version(3)
-            .build(ctx.engine.as_ref())?;
-        // snapshot_v3 was built via log replay, which correctly picks up the v2 metadata
-        // change (pruned replay over commits after CRC@v1 finds the new metaData action).
-        assert_eq!(
-            snapshot_v3
-                .table_properties()
-                .unknown_properties
-                .get("delta.myKey"),
-            Some(&"myValue".to_string())
-        );
-
-        // Write checkpoint@v2, incorporating the new metadata.
-        Snapshot::builder_for(table_root)
-            .at_version(2)
-            .build(ctx.engine.as_ref())?
-            .checkpoint(ctx.engine.as_ref(), None)?;
-
-        // Incremental update. With the filter: CRC@v1 dropped, log replay via checkpoint@v2
-        // returns the correct new metadata. Without the filter: Case 2(b) fallback returns
-        // CRC@v1's stale (empty) configuration, and compare_snapshots fails on
-        // table_configuration.
-        let updated = Snapshot::builder_from(snapshot_v3).build(ctx.engine.as_ref())?;
-        let fresh = Snapshot::builder_for(table_root).build(ctx.engine.as_ref())?;
-        compare_snapshots(&updated, &fresh);
-
-        // Observable contract: the user-visible metadata reflects the v2 change, not the
-        // v0 baseline captured by the stale CRC.
-        assert_eq!(
-            updated
-                .table_properties()
-                .unknown_properties
-                .get("delta.myKey"),
-            Some(&"myValue".to_string())
-        );
-
-        Ok(())
-    }
-
-    // Setup: commits 0-3 with a metadata change at commit 2; stale CRC@v1 captures the
-    // pre-change P&M; snapshot_v3 picks up the v2 metadata via log replay. A new
-    // checkpoint@v1 is then written (at or below snapshot_v3, so the update takes the
-    // P+M-replay path). The pruned replay on commits > v3 is empty, so without the skip
-    // it would fall back to CRC@v1 and regress to the pre-v2 configuration.
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_incremental_snapshot_preserves_metadata_when_below_checkpoint_crc_is_stale(
-    ) -> DeltaResult<()> {
-        let ctx = setup_incremental_snapshot_test()?;
-        let table_root = ctx.url.as_str();
-
-        // commit 0: initial P&M with empty configuration.
-        commit(
-            table_root,
-            &ctx.store,
-            0,
-            vec![
-                protocol_action(1, 2),
-                metadata_action(json!({})),
-                add_action("f0.parquet"),
-            ],
-        )
-        .await;
-        // commit 1: data-only add.
-        commit(table_root, &ctx.store, 1, vec![add_action("f1.parquet")]).await;
-        // commit 2: metadata change (new configuration property).
-        commit(
-            table_root,
-            &ctx.store,
-            2,
-            vec![
-                metadata_action(json!({"delta.myKey": "myValue"})),
-                add_action("f2.parquet"),
-            ],
-        )
-        .await;
-        // commit 3: data-only add.
-        commit(table_root, &ctx.store, 3, vec![add_action("f3.parquet")]).await;
-
-        // CRC@v1 captures pre-v2 metadata (empty configuration); "stale" relative to the
-        // existing snapshot's v2-derived P&M.
-        ctx.store
-            .put(
-                &delta_path_for_version(1, "crc"),
-                make_test_crc_json(200, 2).to_string().into(),
-            )
-            .await?;
-
-        let snapshot_v3 = Snapshot::builder_for(table_root)
-            .at_version(3)
-            .build(ctx.engine.as_ref())?;
-        // Log replay picks up v2's metadata change despite CRC@v1's stale snapshot.
-        assert_eq!(
-            snapshot_v3
-                .table_properties()
-                .unknown_properties
-                .get("delta.myKey"),
-            Some(&"myValue".to_string())
-        );
-
-        // Write checkpoint@v1 (at or below snapshot_v3, so incremental update takes the
-        // P+M-replay path instead of a full rebuild).
-        Snapshot::builder_for(table_root)
-            .at_version(1)
-            .build(ctx.engine.as_ref())?
-            .checkpoint(ctx.engine.as_ref(), None)?;
-
-        // Incremental update. The pruned replay on commits > 3 is empty. With the stale-
-        // CRC skip, the replay returns (None, None) and `TableConfiguration::try_new_from`
-        // preserves existing P&M. Without the skip, Case 2(b) would load CRC@v1's stale
-        // metadata and overwrite the existing (correct) v2 configuration.
-        let updated = Snapshot::builder_from(snapshot_v3).build(ctx.engine.as_ref())?;
-        let fresh = Snapshot::builder_for(table_root).build(ctx.engine.as_ref())?;
-        compare_snapshots(&updated, &fresh);
-
-        // Observable contract: v2 metadata preserved.
-        assert_eq!(
-            updated
-                .table_properties()
-                .unknown_properties
-                .get("delta.myKey"),
-            Some(&"myValue".to_string())
-        );
-
-        Ok(())
-    }
-
-    // Multi-hop: v0 -> v5 (incremental P+M replay) -> v10 (checkpoint-ahead full rebuild).
-    // Verifies that a loaded LazyCrc carried into the second hop is *not* reused when
-    // the new checkpoint invalidates the existing CRC (resulting LazyCrc is fresh and
-    // unloaded, not the loaded Arc from snapshot_v5).
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_incremental_snapshot_multi_hop_replay_then_rebuild_drops_stale_crc(
-    ) -> DeltaResult<()> {
-        let ctx = setup_incremental_snapshot_test()?;
-        setup_test_table_with_commits(ctx.url.as_str(), &ctx.store, 11).await?;
-
-        // CRC at v5 with matching P&M; it is well above any checkpoint (none exists yet),
-        // so every hop inherits it.
-        ctx.store
-            .put(
-                &delta_path_for_version(5, "crc"),
-                make_test_crc_json(600, 6).to_string().into(),
-            )
-            .await?;
-
-        // Hop 1: snapshot at v0.
-        let snapshot_v0 = Snapshot::builder_for(ctx.url.as_str())
-            .at_version(0)
-            .build(ctx.engine.as_ref())?;
-        assert_eq!(snapshot_v0.lazy_crc.crc_version(), None);
-
-        // Hop 2: incremental v0 -> v5 (new commits, no checkpoint -> P+M replay). This picks up
-        // CRC@v5 from the listing.
-        let snapshot_v5 = Snapshot::builder_from(snapshot_v0).build(ctx.engine.as_ref())?;
-        assert_eq!(snapshot_v5.log_segment.checkpoint_version, None);
-        assert_eq!(snapshot_v5.lazy_crc.crc_version(), Some(5));
-
-        // Force-load the v5 CRC so we can observe that a dropped CRC yields a fresh
-        // unloaded LazyCrc on the next hop (not the loaded v5 Arc carried forward).
-        let _ = snapshot_v5.lazy_crc.get_or_load(ctx.engine.as_ref());
-        assert!(snapshot_v5.lazy_crc.is_loaded());
-
-        // Write checkpoint at v7.
-        Snapshot::builder_for(ctx.url.as_str())
-            .at_version(7)
-            .build(ctx.engine.as_ref())?
-            .checkpoint(ctx.engine.as_ref(), None)?;
-
-        // Hop 3: incremental v5 -> v10 (discovers a new checkpoint@v7 ahead of the
-        // existing snapshot version, triggering full rebuild). CRC@v5 is below
-        // the new checkpoint (5 < 7) so the filter drops it, and the rebuilt snapshot gets
-        // a fresh unloaded LazyCrc rather than the v5 Arc that was loaded above.
-        let snapshot_v10 = Snapshot::builder_from(snapshot_v5).build(ctx.engine.as_ref())?;
-        let fresh_v10 = Snapshot::builder_for(ctx.url.as_str()).build(ctx.engine.as_ref())?;
-        compare_snapshots(&snapshot_v10, &fresh_v10);
-        assert_eq!(snapshot_v10.version(), 10);
-        assert_eq!(snapshot_v10.log_segment.checkpoint_version, Some(7));
-        assert_eq!(snapshot_v10.lazy_crc.crc_version(), None);
-        assert!(!snapshot_v10.lazy_crc.is_loaded());
-
-        Ok(())
-    }
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_explicit_same_version_request_keeps_existing_snapshot_after_checkpoint_write(
-    ) -> DeltaResult<()> {
-        let ctx = setup_incremental_snapshot_test()?;
-
-        setup_test_table_with_commits(ctx.url.as_str(), &ctx.store, 2).await?;
-
-        let snapshot_v1 = Snapshot::builder_for(ctx.url.as_str())
-            .at_version(1)
-            .build(ctx.engine.as_ref())?;
-        assert_eq!(snapshot_v1.log_segment.checkpoint_version, None);
-
-        snapshot_v1.clone().checkpoint(ctx.engine.as_ref(), None)?;
-
-        let refreshed = Snapshot::builder_for(ctx.url.as_str()).build(ctx.engine.as_ref())?;
-        assert_eq!(refreshed.log_segment.checkpoint_version, Some(1));
-
-        let pinned = Snapshot::builder_from(snapshot_v1.clone())
-            .at_version(1)
-            .build(ctx.engine.as_ref())?;
-        assert!(Arc::ptr_eq(&pinned, &snapshot_v1));
-        assert_eq!(pinned.log_segment.checkpoint_version, None);
-
-        Ok(())
-    }
-
-    // TODO(#2337): remove this test when log compaction is re-enabled.
-    #[tokio::test]
-    async fn test_compaction_files_ignored_on_read() -> DeltaResult<()> {
-        let store = Arc::new(InMemory::new());
-        let table_root = "memory:///";
-        let engine = SyncEngine::new_with_store(store.clone());
-
-        // Create commits 0-2 and write compaction files to storage
-        setup_test_table_with_commits(table_root, &store, 3).await?;
-        write_compaction_file(&store, 0, 1).await?;
-        write_compaction_file(&store, 0, 2).await?;
-
-        // Compaction files exist on disk but should be skipped during listing
-        let snapshot = Snapshot::builder_for(table_root).build(&engine)?;
-        assert!(
-            snapshot
-                .log_segment
-                .listed
-                .ascending_compaction_files
-                .is_empty(),
-            "Compaction files should be ignored when log compaction is disabled"
-        );
-
-        Ok(())
-    }
-
-    // TODO(#2337): remove this test when log compaction is re-enabled.
-    #[tokio::test]
-    async fn test_incremental_snapshot_ignores_compaction_files() -> DeltaResult<()> {
-        let store = Arc::new(InMemory::new());
-        let table_root = "memory:///";
-        let engine = SyncEngine::new_with_store(store.clone());
-
-        // Create commits 0-2 with compaction files in storage
-        setup_test_table_with_commits(table_root, &store, 3).await?;
-        write_compaction_file(&store, 0, 1).await?;
-        write_compaction_file(&store, 0, 2).await?;
-
-        // Build base snapshot at v2
-        let snapshot_v2 = Snapshot::builder_for(table_root)
-            .at_version(2)
-            .build(&engine)?;
-        assert_eq!(snapshot_v2.version(), 2);
-        assert!(snapshot_v2
-            .log_segment
-            .listed
-            .ascending_compaction_files
-            .is_empty());
-
-        // Add commit 3
-        commit(
-            table_root,
-            &store,
-            3,
-            vec![json!({"add": {"path": "file4.parquet", "partitionValues": {}, "size": 400, "modificationTime": 4000, "dataChange": true}})],
-        )
-        .await;
-
-        // Build v3 incrementally -- compaction files should still be skipped
-        let snapshot_v3 = Snapshot::builder_from(snapshot_v2)
-            .at_version(3)
-            .build(&engine)?;
-        assert_eq!(snapshot_v3.version(), 3);
-        assert!(snapshot_v3
-            .log_segment
-            .listed
-            .ascending_compaction_files
-            .is_empty());
-
-        Ok(())
-    }
-
-    /// When a checkpoint at or below the existing snapshot version is discovered and commits
-    /// between that checkpoint and the existing snapshot version contain P&M changes, the
-    /// incremental update must preserve those P&M changes rather than regressing to the
-    /// checkpoint's older P&M.
-    ///
-    /// Setup: initial checkpoint@v0 so listing_start=1. A checkpoint at v1 is written (discoverable
-    /// since v1 >= listing_start). The P&M change at commit 2 lies between the new checkpoint (v1)
-    /// and existing_snapshot_version (v3); the incremental update must preserve it.
-    #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-    async fn test_checkpoint_at_or_below_snapshot_version_preserves_pm_from_commits_in_between(
-    ) -> DeltaResult<()> {
-        let ctx = setup_incremental_snapshot_test()?;
-        let table_root = ctx.url.as_str();
-
-        // Commit 0: initial protocol (1, 2) + metadata + first file.
-        commit(
-            table_root,
-            &ctx.store,
-            0,
-            vec![
-                protocol_action(1, 2),
-                metadata_action(json!({})),
-                add_action("file1.parquet"),
-            ],
-        )
-        .await;
-        // Commit 1: data-only add.
-        commit(table_root, &ctx.store, 1, vec![add_action("file2.parquet")]).await;
-        // Commit 2: protocol upgrade to (2, 5). This is the change the incremental update
-        // must preserve rather than regressing to checkpoint@v1's older protocol.
-        commit(table_root, &ctx.store, 2, vec![protocol_action(2, 5)]).await;
-        // Commit 3: data-only add.
-        commit(table_root, &ctx.store, 3, vec![add_action("file3.parquet")]).await;
-
-        // Write initial checkpoint at v0 so that listing_start = v0+1 = 1.
-        Snapshot::builder_for(table_root)
-            .at_version(0)
-            .build(ctx.engine.as_ref())?
-            .checkpoint(ctx.engine.as_ref(), None)?;
-
-        // Build snapshot at v3: has checkpoint@v0, protocol (2, 5) from commit 2.
-        let snapshot_v3 = Snapshot::builder_for(table_root)
-            .at_version(3)
-            .build(ctx.engine.as_ref())?;
-        assert_eq!(snapshot_v3.log_segment.checkpoint_version, Some(0));
-        assert_eq!(
-            snapshot_v3
-                .table_configuration()
-                .protocol()
-                .min_reader_version(),
-            2
-        );
-        assert_eq!(
-            snapshot_v3
-                .table_configuration()
-                .protocol()
-                .min_writer_version(),
-            5
-        );
-
-        // Write a checkpoint at v1 (v1 >= listing_start=1, so it is discoverable).
-        // This checkpoint captures only protocol (1, 2) from commit 0.
-        Snapshot::builder_for(table_root)
-            .at_version(1)
-            .build(ctx.engine.as_ref())?
-            .checkpoint(ctx.engine.as_ref(), None)?;
-
-        // Commit 4: add file (new state beyond existing_snapshot_version=3).
-        commit(table_root, &ctx.store, 4, vec![add_action("file4.parquet")]).await;
-
-        // Incremental update from v3: listing starts from v1, discovers checkpoint@v1.
-        // Must preserve protocol (2, 5) from commit 2, not regress to (1, 2) from checkpoint@v1.
-        let updated = Snapshot::builder_from(snapshot_v3).build(ctx.engine.as_ref())?;
-        assert_eq!(updated.version(), 4);
-        assert_eq!(updated.log_segment.checkpoint_version, Some(1));
-
-        // Critical: protocol must still be (2, 5) from commit 2, not (1, 2) from ckpt@v1.
-        assert_eq!(
-            updated
-                .table_configuration()
-                .protocol()
-                .min_reader_version(),
-            2
-        );
-        assert_eq!(
-            updated
-                .table_configuration()
-                .protocol()
-                .min_writer_version(),
-            5
-        );
-
-        // Cross-check against a fresh build.
-        let fresh = Snapshot::builder_for(ctx.url.as_str()).build(ctx.engine.as_ref())?;
-        compare_snapshots(&updated, &fresh);
-
-        Ok(())
-    }
-
-    /// The incremental snapshot path (try_new_from_impl) re-lists files from the checkpoint
-    /// version onwards. We must ensure that it deduplicates compaction files, since producing
-    /// duplicates violated the sort invariant in LogSegmentFilesBuilder::build().
-    #[tokio::test]
-    #[ignore = "log compaction disabled (#2337)"]
-    async fn test_incremental_snapshot_with_compaction_files() -> DeltaResult<()> {
-        let store = Arc::new(InMemory::new());
-        let table_root = "memory:///";
-        let engine = SyncEngine::new_with_store(store.clone());
-
-        // Create commits 0-3 and compaction files (1,1) and (1,2)
-        setup_test_table_with_commits(table_root, &store, 3).await?;
-        write_compaction_file(&store, 1, 1).await?;
-        write_compaction_file(&store, 1, 2).await?;
-
-        // Build snapshot at v2 (includes both compaction files)
-        let snapshot_v2 = Snapshot::builder_for(table_root)
-            .at_version(2)
-            .build(&engine)?;
-        assert_eq!(
-            snapshot_v2
-                .log_segment
-                .listed
-                .ascending_compaction_files
-                .len(),
-            2
-        );
-
-        // Add commit 3
-        commit(
-            table_root,
-            &store,
-            3,
-            vec![json!({"add": {"path": "file4.parquet", "partitionValues": {}, "size": 400, "modificationTime": 4000, "dataChange": true}})],
-        )
-        .await;
-
-        // Build v3 incrementally - before the fix, this panicked due to duplicate compaction files
-        let snapshot_v3 = Snapshot::builder_from(snapshot_v2)
-            .at_version(3)
-            .build(&engine)?;
-
-        assert_eq!(snapshot_v3.version(), 3);
-        assert_eq!(
-            snapshot_v3
-                .log_segment
-                .listed
-                .ascending_compaction_files
-                .len(),
-            2
-        );
-
-        Ok(())
-    }
-
-    /// This test documents a limitation: When deduplicating compactions, the deduplication logic
-    /// only checks the start version (lo), not the hi version. So a new compaction file (1,3)
-    /// added after building the base snapshot at v2 gets filtered out because its start version
-    /// (1) <= existing_snapshot_version (2).
-    #[tokio::test]
-    #[ignore = "log compaction disabled (#2337)"]
-    async fn test_incremental_snapshot_with_new_compaction_files() -> DeltaResult<()> {
-        let store = Arc::new(InMemory::new());
-        let table_root = "memory:///";
-        let engine = SyncEngine::new_with_store(store.clone());
-
-        // Create commits 0-3 and compaction files (1,2) and (2,2)
-        setup_test_table_with_commits(table_root, &store, 4).await?;
-        write_compaction_file(&store, 1, 2).await?;
-        write_compaction_file(&store, 2, 2).await?;
-
-        // Build snapshot at v2
-        let snapshot_v2 = Snapshot::builder_for(table_root)
-            .at_version(2)
-            .build(&engine)?;
-        assert_eq!(
-            snapshot_v2
-                .log_segment
-                .listed
-                .ascending_compaction_files
-                .len(),
-            2
-        );
-
-        // Add new compaction file (1,3) after building the base snapshot
-        write_compaction_file(&store, 1, 3).await?;
-
-        // Build v3 incrementally - the new (1,3) file gets filtered out because
-        // the deduplication only looks at start version: 1 <= existing_snapshot_version (2)
-        let snapshot_v3 = Snapshot::builder_from(snapshot_v2)
-            .at_version(3)
-            .build(&engine)?;
-
-        assert_eq!(snapshot_v3.version(), 3);
-        assert_eq!(
-            snapshot_v3
-                .log_segment
-                .listed
-                .ascending_compaction_files
-                .len(),
-            2
-        );
-
-        // Verify we still have the original (1,2) and (2,2) files
-        let versions_and_his: Vec<_> = snapshot_v3
-            .log_segment
-            .listed
-            .ascending_compaction_files
-            .iter()
-            .map(|p| match p.file_type {
-                LogPathFileType::CompactedCommit { hi } => (p.version, hi),
-                _ => panic!("Expected CompactedCommit"),
-            })
-            .collect();
-        assert_eq!(versions_and_his, vec![(1, 2), (2, 2)]);
-
-        Ok(())
     }
 
     #[test]
